Support detecting out-of-sync pins and more logs in ktsctl.

+++ ktsctl/src/main.rs
@@ -11,13 +11,21 @@ use std::{
 use clap::Parser;
 use cli::Cli;
 use colored::Colorize;
-use kak_tree_sitter_config::{source::Source, Config, LanguageConfig};
+use kak_tree_sitter_config::{
+  source::Source, Config, LanguageConfig, LanguageGrammarConfig, LanguageQueriesConfig,
+};
 use thiserror::Error;
 
 mod cli;
 
 #[derive(Debug, Error)]
 pub enum AppError {
+  #[error("logger failed to initialize: {err}")]
+  LoggerError {
+    #[from]
+    err: log::SetLoggerError,
+  },
+
   #[error("no runtime directory available")]
   NoRuntimeDir,
 
@@ -164,7 +172,13 @@ fn kak_tree_sitter_data_dir() -> Result<PathBuf, AppError> {
 
 fn start() -> Result<(), AppError> {
   let cli = Cli::parse();
+
+  if cli.verbose {
+    simple_logger::init_with_level(log::Level::Debug)?;
+  }
+
   let config = Config::load_default_user()?;
+  log::debug!("ktsctl configuration:\n{config:#?}");
 
   // check the runtime dir exists
   let runtime_dir = runtime_dir()?;
@@ -415,7 +429,7 @@ fn display_lang_info(config: &Config, install_dir: &Path, lang: &str) -> Result<
   };
 
   display_lang_config(lang_config);
-  display_lang_install_stats(install_dir, lang);
+  display_lang_install_stats(lang_config, install_dir, lang);
 
   Ok(())
 }
@@ -543,34 +557,58 @@ fn no_sign() -> impl Display {
   "".red()
 }
 
-fn display_lang_install_stats(install_dir: &Path, lang: &str) {
+fn display_lang_install_stats(lang_config: &LanguageConfig, install_dir: &Path, lang: &str) {
   println!("{section}", section = config_section("Install stats"));
 
-  let grammar_path = install_dir.join(format!("grammars/{lang}.so"));
-  if let Ok(true) = grammar_path.try_exists() {
-    println!(
-      "   {sign} {grammar}{delim} {path}",
-      sign = check_sign(),
-      grammar = format!("{lang} grammar").blue(),
-      delim = delim(":"),
-      path = format!("{}", grammar_path.display()).green()
-    );
-  } else {
-    println!(
-      "   {sign} {lang} grammar missing; install with {help}",
-      sign = no_sign(),
-      help = format!("ktsctl manage -fci {lang}").bold()
-    );
+  display_grammar_info(&lang_config.grammar, install_dir, lang);
+  display_queries_info(&lang_config.queries, install_dir, lang);
+}
+
+/// Check install grammar and report status.
+fn display_grammar_info(config: &LanguageGrammarConfig, install_dir: &Path, lang: &str) {
+  let grammar_path = match config.source {
+    Source::Local { ref path } => path.clone(),
+    Source::Git { ref pin, .. } => install_dir.join(format!("grammars/{lang}/{pin}.so")),
   };
 
-  let queries_path = install_dir.join(format!("queries/{lang}"));
-  display_queries_info(&queries_path, lang);
+  if let Ok(true) = grammar_path.try_exists() {
+    println!(
+      "   {sign} {grammar}",
+      sign = check_sign(),
+      grammar = format!("{lang} grammar").blue(),
+    );
+  } else {
+    let grammars_path = install_dir.join(format!("grammars/{lang}"));
+
+    if let Ok(true) = grammars_path.try_exists() {
+      // we might have a list of stuff
+      println!(
+        "   {sign} {lang} grammar out of sync; synchronize with {help}",
+        sign = no_sign(),
+        help = format!("ktsctl manage --sync --lang {lang}").bold()
+      );
+    } else {
+      println!(
+        "   {sign} {lang} grammar missing; install with {help}",
+        sign = no_sign(),
+        help = format!("ktsctl manage -fci {lang}").bold()
+      );
+    }
+  };
 }
 
 /// Check installed queries and report status.
-fn display_queries_info(path: &Path, lang: &str) {
-  if let Ok(true) = path.try_exists() {
-    let scm_files: HashSet<_> = path
+fn display_queries_info(config: &LanguageQueriesConfig, install_dir: &Path, lang: &str) {
+  let Some(source) = config.source.as_ref() else {
+    return;
+  };
+  let queries_path = match source {
+    Source::Local { path } => path.clone(),
+    Source::Git { pin, .. } => install_dir.join(format!("queries/{lang}/{pin}")),
+  };
+
+  if let Ok(true) = queries_path.try_exists() {
+    let scm_files: HashSet<_> = queries_path
       .read_dir()
       .into_iter()
       .flatten()
@@ -605,19 +643,15 @@ fn display_queries_info(path: &Path, lang: &str) {
 
     if scm_count == scm_expected_count {
       println!(
-        "   {sign} {queries}{delim} {path}",
+        "   {sign} {queries}",
         sign = check_sign(),
         queries = format!("{lang} queries").blue(),
-        delim = delim(":"),
-        path = format!("{}", path.display()).green(),
       );
     } else if scm_count > 0 {
       println!(
-        "   {sign} {queries}{delim} {path}",
+        "   {sign} {queries}",
         sign = warn_sign(),
         queries = format!("{lang} queries").blue(),
-        delim = delim(":"),
-        path = path.display()
       );
     } else {
       println!(
@@ -631,11 +665,21 @@ fn display_queries_info(path: &Path, lang: &str) {
       println!("{q}");
     }
   } else {
-    println!(
-      "   {sign} {lang} queries missing; install with {help}",
-      sign = no_sign(),
-      help = format!("ktsctl manage -fci {lang}").bold()
-    );
+    let queries_path = install_dir.join(format!("queries/{lang}"));
+
+    if let Ok(true) = queries_path.try_exists() {
+      println!(
+        "   {sync} {lang} queries out of sync; synchronize with {help}",
+        sync = no_sign(),
+        help = format!("ktsctl manage --sync {lang}").bold()
+      );
+    } else {
+      println!(
+        "   {sign} {lang} queries missing; install with {help}",
+        sign = no_sign(),
+        help = format!("ktsctl manage -fci {lang}").bold()
+      );
+    }
   }
 }
